+++ src/Validation/Filter/Type.php
@@ -11,6 +11,8 @@ use Message\Cog\Validation\Loader;
  *
  * Casts fields to specific data types.
  *
+ * @deprecated Do not use this component, use Symfony's validation component instead
+ *
  * @author James Moss <james@message.co.uk>
  * @author Thomas Marchant <thomas@message.co.uk>
  */